+++ src/org/trifort/rootbeer/runtime/ThreadConfig.java
@@ -11,9 +11,9 @@ public class ThreadConfig {
 
   private int m_blockShapeX;
   private int m_gridShapeX;
-  private int m_numThreads;
+  private long m_numThreads;
   
-  public ThreadConfig(int block_shape_x, int grid_shape_x,int numThreads){
+  public ThreadConfig(int block_shape_x, int grid_shape_x, long numThreads){
     m_blockShapeX = block_shape_x;
     m_gridShapeX = grid_shape_x;
     m_numThreads = numThreads;
@@ -27,7 +27,7 @@ public class ThreadConfig {
     return m_gridShapeX;
   }
   
-  public int getNumThreads(){
+  public long getNumThreads(){
     return m_numThreads;
   }
